% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_color_rki}
\alias{scale_color_rki}
\title{RKI color scale for ggplot2 with different pallettes}
\usage{
scale_color_rki(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{Options}{include palette (Defaults to "main" RKI Blue-Shades, with other options beeing "all", "qualitative")}
}
\description{
RKI color scale for ggplot2 with different pallettes
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  theme_rki() + scale_color_rki()


}
\keyword{custom_scale_color()}
