% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_fill_rki}
\alias{scale_fill_rki}
\title{RKI fill scale for ggplot2 with different pallettes}
\usage{
scale_fill_rki(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{Options}{include palette (Defaults to "main" RKI Blue-Shades, with other options beeing "all", "qualitative")}
}
\description{
RKI fill scale for ggplot2 with different pallettes
}
\examples{
ggplot(iris, aes(x=Species,y=Sepal.Width, fill = Species)) +   geom_bar(stat = "summary", fun.y = "mean") +  theme_rki() + scale_fill_rki(


}
\keyword{custom_scale_color()}
