% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_rki}
\alias{theme_rki}
\title{RKI Theme based on theme_grey}
\usage{
theme_rki(base_size = 12, base_family = "", bold_family = "")
}
\arguments{
\item{Define}{a base_size (Defaults to 12) and base_family and bold_family for Fonts used (defaults to ggplot2's defaults)}
}
\description{
RKI Theme based on theme_grey
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  theme_rki()

}
\keyword{theme}
