% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_rki_void}
\alias{theme_rki_void}
\title{RKI Theme without axis and labels based on theme_rki() optimal suited to print maps.}
\usage{
theme_rki_void(base_size = 12, base_family = "")
}
\arguments{
\item{Define}{a base_size (Defaults to 12) and base_family for Fonts used (defaults to ggplot2's defaults)}
}
\description{
RKI Theme without axis and labels based on theme_rki() optimal suited to print maps.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  theme_rki_void()

}
\keyword{theme}
